Create namespaces for enemies
