--- Assets/Scripts/Enemies/EnemyContactDamage.cs
@@ -1,18 +1,0 @@
-using System;
-using System.Collections;
-using System.Collections.Generic;
-using Unity.VisualScripting;
-using UnityEngine;
-
-public class EnemyContactDamage : MonoBehaviour
-{
-    [SerializeField] private int contactDamage = 1;
-
-    private void OnTriggerEnter2D(Collider2D col)
-    {
-        if (col.gameObject.CompareTag("Player"))
-        {
-            col.gameObject.GetComponent<PlayerHealth>().TakeDamage(contactDamage);
-        }
-    }
-}
