+++ Assets/Scripts/Enemies/ParabolicRangedAttack.cs.meta
