+++ Assets/Sprites 1.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1bb2b912757281948b4478e9efeab9c8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
